+++ library/src/androidTest/java/InboxTests.kt
@@ -1,7 +1,7 @@
 package com.wultra.android.mtokensdk.test
 
 import com.wultra.android.mtokensdk.inbox.*
-import com.wultra.android.mtokensdk.operation.IOperationsService
+import com.wultra.android.mtokensdk.operation.OperationsService
 import io.getlime.security.powerauth.sdk.PowerAuthSDK
 import org.junit.*
 import org.junit.Assert.*
@@ -10,8 +10,8 @@ import java.util.concurrent.TimeUnit
 
 class InboxTests {
 
-    lateinit var ops: IOperationsService
-    lateinit var inbox: IInboxService
+    lateinit var ops: OperationsService
+    lateinit var inbox: InboxService
     private lateinit var pa: PowerAuthSDK
     val pin = "1234"
 
@@ -19,9 +19,9 @@ class InboxTests {
     fun setup() {
         try {
             val result = IntegrationUtils.prepareActivation(pin)
-            pa = result.pa
-            ops = result.ops
-            inbox = result.inbox
+            pa = result.first
+            ops = result.second.operations
+            inbox = result.second.inbox
         } catch (e: Throwable) {
             fail("Activation preparation failed: $e")
         }