+++ library/src/main/java/com/wultra/android/mtokensdk/inbox/InboxService.kt
@@ -17,6 +17,7 @@
 package com.wultra.android.mtokensdk.inbox
 
 import android.content.Context
+import com.google.gson.GsonBuilder
 import com.wultra.android.mtokensdk.api.inbox.*
 import com.wultra.android.mtokensdk.api.inbox.model.GetList
 import com.wultra.android.mtokensdk.api.inbox.model.GetMessageDetail
@@ -28,66 +29,58 @@ import com.wultra.android.powerauth.networking.UserAgent
 import com.wultra.android.powerauth.networking.data.StatusResponse
 import com.wultra.android.powerauth.networking.error.ApiError
 import com.wultra.android.powerauth.networking.error.ApiErrorException
-import com.wultra.android.powerauth.networking.ssl.SSLValidationStrategy
 import com.wultra.android.powerauth.networking.tokens.IPowerAuthTokenProvider
 import io.getlime.security.powerauth.sdk.PowerAuthSDK
 import okhttp3.OkHttpClient
 
 /**
- * Convenience factory method to create an IInboxService instance
- * from given PowerAuthSDK instance.
- *
- * @param appContext Application Context
- * @param baseURL Base URL for inbox request  (ending with `/enrollment-server` in the default setup)
- * @param okHttpClient HTTP client instance for networking
- * @param userAgent Default user agent for each request.
- * @return IInboxService instance
+ * Service that communicates with Inbox API that is managing user's inbox.
  */
-fun PowerAuthSDK.createInboxService(appContext: Context, baseURL: String, okHttpClient: OkHttpClient, userAgent: UserAgent? = null): IInboxService {
-    return InboxService(okHttpClient, baseURL, this, appContext, null, userAgent)
-}
-
-/**
- * Convenience factory method to create an IInboxService instance
- * from given PowerAuthSDK instance.
- *
- * @param appContext Application Context
- * @param baseURL Base URL for inbox request  (ending with `/enrollment-server` in the default setup)
- * @param strategy SSL validation strategy for networking
- * @param userAgent Default user agent for each request.
- * @return IInboxService instance
- */
-fun PowerAuthSDK.createInboxService(appContext: Context, baseURL: String, strategy: SSLValidationStrategy, userAgent: UserAgent? = null): IInboxService {
-    val builder = OkHttpClient.Builder()
-    strategy.configure(builder)
-    return InboxService(builder.build(), baseURL, this, appContext, null, userAgent)
-}
-
 class InboxService(
-    httpClient: OkHttpClient,
-    baseURL: String,
     powerAuthSDK: PowerAuthSDK,
     appContext: Context,
+    httpClient: OkHttpClient,
+    baseURL: String,
     tokenProvider: IPowerAuthTokenProvider? = null,
-    userAgent: UserAgent? = null
-) : IInboxService {
+    userAgent: UserAgent? = null,
+    gsonBuilder: GsonBuilder? = null
+) {
 
     // API class for communication.
-    private val inboxApi = InboxApi(httpClient, baseURL, appContext, powerAuthSDK, tokenProvider, userAgent, OperationsUtils.defaultGsonBuilder())
+    private val inboxApi = InboxApi(httpClient, baseURL, appContext, powerAuthSDK, tokenProvider, userAgent, gsonBuilder ?: OperationsUtils.defaultGsonBuilder())
 
-    override var acceptLanguage: String
+    /**
+     * Accept language for the outgoing requests headers.
+     * Default value is "en".
+     * Changing this value updates the accept language of the underlying inboxApi.
+     *
+     * Standard RFC "Accept-Language" https://tools.ietf.org/html/rfc7231#section-5.3.5
+     * Response texts are based on this setting. For example when "de" is set, server
+     * will return operation texts in german (if available).
+     */
+    var acceptLanguage: String
         get() = inboxApi.acceptLanguage
         set(value) {
             inboxApi.acceptLanguage = value
         }
 
-    override var okHttpInterceptor: OkHttpBuilderInterceptor?
+    /**
+     * A custom interceptor can intercept each service call.
+     *
+     * You can use this for request/response logging into your own log system.
+     */
+    var okHttpInterceptor: OkHttpBuilderInterceptor?
         get() = inboxApi.okHttpInterceptor
         set(value) {
             inboxApi.okHttpInterceptor = value
         }
 
-    override fun getUnreadCount(callback: (result: Result<InboxCount>) -> Unit) {
+    /**
+     * Get number of unread messages in the inbox.
+     *
+     * @param callback Callback with result.
+     */
+    fun getUnreadCount(callback: (result: Result<InboxCount>) -> Unit) {
         inboxApi.count(object: IApiCallResponseListener<InboxCountResponse> {
             override fun onFailure(error: ApiError) {
                 callback(Result.failure(ApiErrorException(error)))
@@ -99,7 +92,15 @@ class InboxService(
         })
     }
 
-    override fun getMessageList(pageNumber: Int, pageSize: Int, onlyUnread: Boolean, callback: (result: Result<List<InboxMessage>>) -> Unit) {
+    /**
+     * Paged list of messages in the inbox. You can use also  [getAllMessages] method to fetch all messages.
+     *
+     * @param pageNumber Page number. First page is `0`, second `1`, etc.
+     * @param pageSize Number of items received in the page.
+     * @param onlyUnread If `true` then only unread messages will be returned.
+     * @param callback Result callback. If the number of items in result is less than [pageSize] then the received page is the last page.
+     */
+    fun getMessageList(pageNumber: Int, pageSize: Int, onlyUnread: Boolean, callback: (result: Result<List<InboxMessage>>) -> Unit) {
         inboxApi.list(
             InboxGetListRequest(GetList(pageNumber, pageSize, onlyUnread)),
             object: IApiCallResponseListener<InboxGetListResponse> {
@@ -114,7 +115,13 @@ class InboxService(
         )
     }
 
-    override fun getMessageDetail(messageId: String, callback: (result: Result<InboxMessageDetail>) -> Unit) {
+    /**
+     * Get message detail in the inbox.
+     *
+     * @param messageId Message identifier.
+     * @param callback Result callback.
+     */
+    fun getMessageDetail(messageId: String, callback: (result: Result<InboxMessageDetail>) -> Unit) {
         inboxApi.detail(
             InboxGetMessageDetailRequest(GetMessageDetail(messageId)),
             object: IApiCallResponseListener<InboxGetMessageDetailResponse> {
@@ -129,7 +136,13 @@ class InboxService(
         )
     }
 
-    override fun markRead(messageId: String, callback: (result: Result<Unit>) -> Unit) {
+    /**
+     * Mark the message with the given identifier as read.
+     *
+     * @param messageId Message identifier.
+     * @param callback Result callback.
+     */
+    fun markRead(messageId: String, callback: (result: Result<Unit>) -> Unit) {
         inboxApi.read(
             InboxSetMessageReadRequest(SetMessageRead(messageId)),
             object: IApiCallResponseListener<StatusResponse> {
@@ -144,7 +157,12 @@ class InboxService(
         )
     }
 
-    override fun markAllRead(callback: (result: Result<Unit>) -> Unit) {
+    /**
+     * Mark all unread messages in the inbox as read.
+     *
+     * @param callback Result callback.
+     */
+    fun markAllRead(callback: (result: Result<Unit>) -> Unit) {
         inboxApi.readAll(object : IApiCallResponseListener<StatusResponse> {
             override fun onFailure(error: ApiError) {
                 callback(Result.failure(ApiErrorException(error)))
@@ -154,5 +172,65 @@ class InboxService(
                 callback(Result.success(Unit))
             }
         })
+    }
+
+    /**
+     * Get all messages in the inbox. The function will issue multiple HTTP requests  until the list is not complete.
+     *
+     * @param onlyUnread If `true` then only unread messages will be returned. The default value is `false`.
+     * @param pageSize How many messages should be fetched at once. The default value is 100.
+     * @param messageLimit Maximum number of messages to be retrieved. Use `0` to set no limit. The default value is `1000`.
+     * @param callback Callback with result.
+     */
+    fun getAllMessages(onlyUnread: Boolean = false, pageSize: Int = 100, messageLimit: Int = 1000, callback: (result: Result<List<InboxMessage>>) -> Unit) {
+        fetchPartialList(FetchOperation(0, pageSize, messageLimit, onlyUnread, callback))
+    }
+
+    /**
+     * Fetch partial list from the server.
+     * @param operation [FetchOperation] object.
+     */
+    private fun fetchPartialList(operation: FetchOperation) {
+        getMessageList(operation.pageNumber, operation.pageSize, operation.onlyUnread) { result ->
+            result.onSuccess {
+                if (operation.appendPartialMessages(it)) {
+                    operation.complete()
+                } else {
+                    fetchPartialList(operation)
+                }
+            }.onFailure { operation.complete(result) }
+        }
+    }
+}
+
+/**
+ * Support class that keeps partially received messages and track the current page to fetch.
+ */
+private class FetchOperation(
+    var pageNumber: Int,
+    val pageSize: Int,
+    val messageLimit: Int,
+    val onlyUnread: Boolean,
+    val completion: (Result<List<InboxMessage>>) -> Unit
+) {
+    val readMessages = mutableListOf<InboxMessage>()
+
+    /**
+     * Append received messages and determine whether we're at the end of the list.
+     * @param messages Partial messages received from the server.
+     * @return `true` if we're at the end of the list.
+     */
+    fun appendPartialMessages(messages: List<InboxMessage>): Boolean {
+        readMessages.addAll(messages)
+        pageNumber += 1
+        return messages.size < pageSize || (messageLimit > 0 && readMessages.size >= messageLimit)
+    }
+
+    /**
+     * Complete operation with result.
+     * @param result Result to report back to the application. If `null` then success is reported.
+     */
+    fun complete(result: Result<List<InboxMessage>>? = null) {
+        completion(result ?: Result.success(readMessages))
     }
 }