+++ library/src/main/java/com/wultra/android/mtokensdk/push/PushService.kt
@@ -27,50 +27,36 @@ import com.wultra.android.powerauth.networking.UserAgent
 import com.wultra.android.powerauth.networking.data.StatusResponse
 import com.wultra.android.powerauth.networking.error.ApiError
 import com.wultra.android.powerauth.networking.error.ApiErrorException
-import com.wultra.android.powerauth.networking.ssl.SSLValidationStrategy
 import com.wultra.android.powerauth.networking.tokens.IPowerAuthTokenProvider
 import io.getlime.security.powerauth.sdk.PowerAuthSDK
 import okhttp3.OkHttpClient
 
 /**
- * Convenience factory method to create an IPushService instance
- * from given PowerAuthSDK instance.
- *
- * @param appContext Application Context
- * @param baseURL Base URL for push request  (ending with `/enrollment-server` in the default setup)
- * @param okHttpClient HTTP client instance for networking
- * @param userAgent Default user agent for each request.
- * @return IPushService instance
+ * Service, that communicates with Mobile Token API that handles registration for
+ * push notifications.
  */
-fun PowerAuthSDK.createPushService(appContext: Context, baseURL: String, okHttpClient: OkHttpClient, userAgent: UserAgent? = null): IPushService {
-    return PushService(okHttpClient, baseURL, this, appContext, null, userAgent)
-}
-
-/**
- * Convenience factory method to create an IPushService instance
- * from given PowerAuthSDK instance.
- *
- * @param appContext Application Context
- * @param baseURL Base URL for push request  (ending with `/enrollment-server` in the default setup)
- * @param strategy SSL validation strategy for networking
- * @param userAgent Default user agent for each request.
- * @return IPushService instance
- */
-fun PowerAuthSDK.createPushService(appContext: Context, baseURL: String, strategy: SSLValidationStrategy, userAgent: UserAgent? = null): IPushService {
-    val builder = OkHttpClient.Builder()
-    strategy.configure(builder)
-    return createPushService(appContext, baseURL, builder.build(), userAgent)
-}
-
-class PushService(okHttpClient: OkHttpClient, baseURL: String, powerAuthSDK: PowerAuthSDK, appContext: Context, tokenProvider: IPowerAuthTokenProvider? = null, userAgent: UserAgent? = null): IPushService {
-
-    override var acceptLanguage: String
+class PushService(powerAuthSDK: PowerAuthSDK, appContext: Context, okHttpClient: OkHttpClient, baseURL: String, tokenProvider: IPowerAuthTokenProvider? = null, userAgent: UserAgent? = null) {
+    /**
+     * Accept language for the outgoing requests headers.
+     * Default value is "en".
+     * Changing this value updates the accept language of the underlying pushApi.
+     *
+     * Standard RFC "Accept-Language" https://tools.ietf.org/html/rfc7231#section-5.3.5
+     * Response texts are based on this setting. For example when "de" is set, server
+     * will return operation texts in german (if available).
+     */
+    var acceptLanguage: String
         get() = pushApi.acceptLanguage
         set(value) {
             pushApi.acceptLanguage = value
         }
 
-    override var okHttpInterceptor: OkHttpBuilderInterceptor?
+    /**
+     * A custom interceptor can intercept each service call.
+     *
+     * You can use this for request/response logging into your own log system.
+     */
+    var okHttpInterceptor: OkHttpBuilderInterceptor?
         get() = pushApi.okHttpInterceptor
         set(value) {
             pushApi.okHttpInterceptor = value
@@ -78,7 +64,13 @@ class PushService(okHttpClient: OkHttpClient, baseURL: String, powerAuthSDK: Pow
 
     private val pushApi = PushApi(okHttpClient, baseURL, powerAuthSDK, appContext, tokenProvider, userAgent)
 
-    override fun register(data: PushData, callback: (result: Result<Unit>) -> Unit) {
+    /**
+     * Registers FCM on PowerAuth backend to receive notifications about operations and inbox.
+     *
+     * @param data Token and push platform (Firebase Cloud Messaging or Huawei Push)
+     * @param callback Result listener
+     */
+    fun register(data: PushData, callback: (result: Result<Unit>) -> Unit) {
         val platform = when (data.platform) {
             PushPlatform.FCM -> PushRegistrationRequestObject.Platform.FCM
             PushPlatform.HMS -> PushRegistrationRequestObject.Platform.HMS
@@ -104,14 +96,13 @@ class PushService(okHttpClient: OkHttpClient, baseURL: String, powerAuthSDK: Pow
     }
 
     // deprecated API
-
     @Deprecated("Use register function with `data` parameter as a replacement")
-    override fun register(fcmToken: String, callback: (Result<Unit>) -> Unit) {
+    fun register(fcmToken: String, callback: (Result<Unit>) -> Unit) {
         register(fcmToken, PushRegistrationRequestObject.Platform.ANDROID, callback)
     }
 
     @Deprecated("Use register function with `data` parameter as a replacement")
-    override fun registerHuawei(hmsToken: String, callback: (result: Result<Unit>) -> Unit) {
+    fun registerHuawei(hmsToken: String, callback: (result: Result<Unit>) -> Unit) {
         register(hmsToken, PushRegistrationRequestObject.Platform.HUAWEI, callback)
     }
 }