+++ docs/Changelog.md
@@ -2,7 +2,7 @@
 
 ## 2.1.0 (Draft)
 
-- Added `mobileTokenData` to authorize request for passing customer-specific data
+- Added `mobileTokenData` to authorize request for passing customer-specific data ([documentation](Using-Operations-Service.md#Passing-Additional-Mobile-Token-Data))
   - Available with PowerAuth server 1.10+
   - Can be used for fraud detection systems (FDS) or other custom business logic